move model capabilities into inference types
